Flask_Blog/requirements.txt
bcrypt==5.0.0
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
click==8.3.1
dotenv==0.9.9
email_validator==2.3.0
Flask==3.1.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Limiter==4.1.1
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
Pillow==12.1.0
pycparser==3.0
pyotp==2.9.0
qrcode[pil]==8.2
six==1.17.0
SQLAlchemy==2.0.46
Werkzeug==3.1.5
WTForms==3.2.1
